Add prefetchInputs calls to native Actions
Some actions in Bazel are implemented in Java and only ever
run inside the Bazel process and never remotely. However,
these actions can have files as inputs that were generated
on a remote system and with #6862 are not stored in the
output base.

I audited these actions manually and added prefetch calls
to the few native actions that didn't have them already.

Progress towards #6862