Windows: Bash is no longer required for everything
Bazel no longer reports an error if it fails to
locate Bash. This allows "bazel info release" to
work even in absence of Bash.

To clarify, Bazel still requires Bash to build
rules that depend on Bash. This PR simply removes
the fatal error Bazel used to display if Bash was
missing even when Bash was not required.

Fixes https://github.com/bazelbuild/bazel/issues/6462