remote: re-implement merkle tree buildling
The TreeNodeRepository served us well, but has become a bit dated over
time and it's increasingly hard to understand and to fix bugs. Additionally,
it was written with the initial intend of incrementally updating the merkle
tree between actions, however internal performance analysis has shown
that doing so is 1) hard to implement reliably and 2) would increase
memory consumption multifold.

This is a rewrite of the code that also fixes bugs like #4663. MerkleTree
implements a visitor pattern over an InputTree and is responsible for
serializing the InputTree to the protobuf merkle tree that's used by
remote caching / execution.

Benchmarks on my local machine have shown this implementation to
consumes between 2-10x less wall time than the current implementation.
Tests by users have shown equally encouraging speedups [1].

[1] https://groups.google.com/d/msg/bazel-discuss/wPHrqm2z8lU/mzoRF236GQAJ

Closes #7583.

PiperOrigin-RevId: 237421145