Teach the FilesystemValueChecker about remotely stored outputs
The FilesystemValueChecker is used by Bazel to detect modified outputs before
running a command. This change teaches it about remote output files that don't
exist in the output base, but are only stored physically on a remote storage
system with SkyFrame only tracking file metadata.

Note that this change only affects incremental builds, and there
introduces the following new behavior for artifacts and tree artifacts:
  1) If the previous build created a remote output and the output
 does not exist in the output base then no invalidation happens.
  2) If the previous build created a remote output and the output
 does exist in the output base then the generating action is
 invalidated.

A tree artifact may either be fully materialized in the output base or
not materialized at all.

Progress towards #6862

Closes #7269.

PiperOrigin-RevId: 237422752