Fix NPE when generating target_complete BEP events for aspects.
We recently added support for populating the configuration field
of a target_complete event generated by an aspect. However, in the
case of source files (and maybe others?), the configuration
can be null. The previous code was not checking for nullness before
trying to get the SkyValue of the BuildConfiguration thus causing an NPE.

PiperOrigin-RevId: 237439110