Propagate Postable events to the ExtendedEventHandlers of Reporter.
Reporter is in charge of handling/posting events to the handlers registered
to it. However, we currently only propagate Reporter#handle calls to the handlers
and not the Reporter#post calls, presumably because posting to the Event bus
was deemed enough. This is actually not true since some of the handlers
in Reporter can be ExtendedEventHandlers which can also post events.

I found this while trying to post an event in beforeCommand and I noticed
that the StoredEventHandler in BlazeCommandDispatcher was replaying only
handle calls but not post calls.

PiperOrigin-RevId: 237452769