Migrate LTO Backends to use features for user and legacy compile flags
Instead of manually composing copt/ltobackendopt and legacy options for LTO
Backend actions, set the user_compile_flags and legacy_compile_flags build
variables. This will allow the compile command line to be set up in the same
manner/order as for the Compile action.

Fixes a problem caused by moving crosstool options into a feature, which was
leading copt/ltobackendopt to be put on the command line before default
options, making it difficult to use them to work around bugs.

RELNOTES:none
PiperOrigin-RevId: 237461044