Transform options by parsing result.
Allows an existing FragmentOptions to be the base for a new one which is distinguished by a set of flags provided through an option parser (for example from the command line or a file).

Step 1/N towards the platforms mapping functionality for https://github.com/bazelbuild/bazel/issues/6426

RELNOTES: None.
PiperOrigin-RevId: 238022069