Fix SpawnAction
Wrapping continuations in more continuations is more complex than it
would seem. The problem is that the code paths to begin execution and to
continue execution ideally perform identical exception handling.
However, if we first call beginExecution and then create a continuation,
we end up duplicating exception handling.

We also have to take into account that beginExecution can return a
finished result - this is the bug fixed here.

Lastly, we ideally never return a continuation that has an execute
method, but does not have a future, i.e., beginExecution should start to
do work before it actually returns. However, this is not a critical
requirement, as SkyframeActionExecutor loops until it sees a non-null
future.

The solution is to create a 'continue' outer continuation wrapping a
'beginExecute' continuation, and call execute exactly once, at the
expense of a small amount of wasted memory.

Progress on #6394.

PiperOrigin-RevId: 238028776