Whitelist Starlark rule class transitions in order to guard roll out.
We already whitelist attribute transitions so a decent amount of testing is done in StarlarkAttrTransitionProviderTest.

Actually checking of the whitelist hinges on having the special whitelist attribute (with proper name and value). So just adding checks that that attribute is properly around when we have rule class transitions hooks in the whitelist checking (see SkylarkRuleConfiguredTargetUtil#buildRule for context).

PiperOrigin-RevId: 238052200