Properly invalidate actions that have tree artifact inputs and do input discovery
While we allowed TreeArtifact header inputs in C++ actions (header validation passed), we failed to invalidate these headers on change. This change fixes it by adding original TreeArtifact as action input when the dotd file reports the header coming from TreeFileArtifact was used during compilation.

A small difference in dotd file handling is that normally .d files are right next to objects, but in case of tree artifacts they will be in a separate directory (_dotd or _pic_dotd).

Fixes #5785

Closes #7452.

PiperOrigin-RevId: 238189242