Add note about limitation of JAVA_HOME paths
Because bazel uses junction points to connect @local_jdk to JAVA_HOME,
and Windows does not support junction points referencing remote
filesystems, it is currently necessary to install Java on a "local"
filesystem.

Using remote_jdk is a partial workaround, but that's more complex and
doesn't belong in this document.

Closes #7711.

PiperOrigin-RevId: 238438245