Allow setting the Blaze server's QoS class via a --macos_qos_class flag.
This adds a new flag to set the QoS class of the Bazel server and accepts
selecting any of the possible classes exposed by the system.  We will use
this flag to evaluate the behavior of Bazel under different classes because
it is not clear which one we should be using.  (We tried forcing the class
to be utility, but that caused regressions in some cases.)

This is essentially a retry of https://github.com/bazelbuild/bazel/commit/087734009801242b83655efb863b2d5a761ae3dc but gated by a flag (which
is how it should have been done in the first place).

Note that the flag exists on all platforms to ensure that a bazelrc file
shared across different people works in all cases, but this flag is ignored
in non-macOS systems.

Addresses https://github.com/bazelbuild/bazel/issues/7446.

RELNOTES: None.
PiperOrigin-RevId: 238440116