Add TestAttemptContinuation
This adds an interface for representing async test attempt execution.
While I'd love to move this up to TestRunnerAction, unfortunately, the
Bazel and Blaze implementations are sufficiently different to make that
difficult at this time.

I have a follow-up change that switches test execution to the new code
for all cases (including non-async), which adds plenty of test coverage.

Progress on #6394.

PiperOrigin-RevId: 238686781