Add an Android Common Starlark API enableImplicitSourcelessDepsExportsCompatibility.
Currently, the native android_library rules is affect by b/14473160. Where the rule will implicitly export its deps when sources are not given in the java compilation. The Starlark android_library will need to implement this behavior. As such, the API is used to convert a JavaInfo normally used as dependency to the java_common.compile, to an "implicit" exportable JavaInfo (meaning that only the JavaCompilationArgsProvider of the JavaInfo is exported).

RELNOTES: None.
PiperOrigin-RevId: 238814381