Generate documentation for embedded repository rules
Our embedded repository rules have doc strings, as recommended. So
far, however, they are not shown on as part of web documentation.
Fix this. The recommended tool for generating web pages from Starlark
documentation strings is Skydoc; therefore, we use it, even if it
pulls in new dependencies, including those that need to be shipped
as part of the distribution archive.

Fixes #6873.
Fixes #7638.
Fixes #6730.

Change-Id: Icb86cf3c321157e692469b6ed0cac9b64334806c
PiperOrigin-RevId: 238958194