bazel protobuf: add AttributeDefinition.cfg_is_host
It records whether an attribute entails a transition to the "host" configuration.

Ideally we would record more information about the transition, but the
interface seems very abstract and the implementations are varied and
complex (arbitrary function logic). The "host" flag is at least part
of the Java interface, part of the Bazel manual, and part of the
Blaze query interface.

PiperOrigin-RevId: 239320132