remote: add support for in memory outputs
in preparation for #6862 add code that allows for Bazel rules to
request outputs in memory. It works by passing allowing an execution
requirement 'internal-inline-outputs' with a path to the output file
whose contents should be provided in memory.

this change introduces some necessary refactorings and adds code
implementing this functionality to the remote package but doesn't
turn this feature on. It will be enabled in a subsequent PR.