Add /Applications/ to cxx_builtin_include_directories
This makes sure that all paths that point inside the Applications
directory on macOS are considered cxx_builtin_include_directories. The
goal here is to make sure that paths in produced `.d` files are
correctly translated by bazel to local paths when the `.d` files are
shared with the remote cache to machines that have Xcode installed in
different locations.