Generate cc_toolchain_cc.bzl instead of CROSSTOOL for android ndks
Before this cl, AndroidNdkRepositoryFunction generated CROSSTOOL file that configured the C++ toolchain for ndk. This cl changes this function to generate a Starlark rule providing CcToolchainConfigInfo, which is the "new" way of configuring C++ toolchains. This change should not be observable by ndk users.

As a result, generated ndk toolchain is forward compatible with incompatible changes:
* https://github.com/bazelbuild/bazel/issues/7008
* https://github.com/bazelbuild/bazel/issues/6861
* https://github.com/bazelbuild/bazel/issues/7320
* https://github.com/bazelbuild/bazel/issues/7075

RELNOTES: Android NDK C++ toolchain is now configured in Starlark. This should be a backwards compatible change, but in case of bugs blame unknown commit.
PiperOrigin-RevId: 239565213