Pass execution info to xml generating spawn. Fixes #7794
Change [1] splits test execution into two spawns. The first spawn runs
the test and the second spawn generates the test.xml from the test.log
file. This change makes it so that the test.xml-generating spawn also
inherits the execution info from the test target. In particular, if a
test target is tagged with "no-remote" then the test.xml-generating
spawn will also be tagged.

[1] https://github.com/bazelbuild/bazel/commit/0858ae1f6eb890c1e203a2aa21130ba34ca36a27

Closes #7796.

PiperOrigin-RevId: 239608851