Move Reporter out of the BuildEventStreamer.
The Reporter is a general-purpose object for communicating events to the user.
When an event is sent (with Reporter#post or Reporter#handle) using the Reporter
then those events are forwarded to the set of registered handlers (for #handle)
and to the event bus/extended handlers for #post. BuildEventStreamer responsibility
is to send BuildEvents to all active transports, it shouldn't
be on the business of sending UI messages (via #handle) or posting to the
event bus (via #post).

With this CL the messages that were sent inside the BuildEventStreamer
are now sent in the BuildEventServiceModule allowing us to manage all the UI
messages from there instead of scattered around the BES/BEP transports code.

From the UI perspective, the ExperimentalStateTracker uses
AnnounceBuildEventTransportsEvent and BuildEventTransportClosedEvent to decide
whether to print information about the BES upload. We now send those events
in the BES module.

In summary, there are a few concrete improvements with this CL:
- The waiting messages are solely the responsibility of the experimental UI.
- The logic for closing the BEP transports in the streamer is simplified.
- The streamer doesn't report arbitrary events to the event handlers or the event bus.

PiperOrigin-RevId: 239613756