Report IllegalArgumentException in URL.openConnection as http error
URL.openConnection can throw an IllegalArgumentException, e.g., in the
JceSecurityManager. While we cannot fix these exceptions, we can at least
report them in a more user-friendly way than an internal stack trace. So
catch and throw an exception that is reported as a permanent error attempting
to fetch the given URL.

Issue #6952.

Change-Id: I2bb5d12535d54046e2cf342acabcb28c2066eed7
PiperOrigin-RevId: 239762567