remote: rename auth flags.
--google_* flags only work with Google Cloud Authentication therefore --gcloud_* prefix would add some clarity for the potential usages
So instead of --google_* let's call them --gcloud_* (the old ones will continue working for a while).

So now there's two simple ways to authenticate with Google Cloud:
* bazel build --gcloud_default_credentials
* bazel build --gcloud_credentials=creds.json

RELNOTES: --google_* flags were renamed to --gcloud_* flags. The old names will continue to work for this release but will be removed in the next release.