Move new UI out of experimental state
The console output stream is turned on by default for quite a while
now, and the decision is made that it should stay the default, and
in the long run the only, console output provided by Bazel. Rename
the respective flags accordingly to avoid giving a wrong impression
about its state.

Change-Id: Iad6240c4f615aae64fc721cd2e26fa5ee30b000d
PiperOrigin-RevId: 239781052