Always track parallelism in EvaluationContext
EvaluationContext has 2 members for different configurations it is run with
(either is set, never both):
  1. Integer numThreads -- this allows delegating the decision of what kind
  ExecutorService to create to the user
  2. Supplier<ExecutorService> executorService -- factory of ExecutorService
  objects -- makes the user tied to what it returns

Some places in the code are interested in knowing the parallelism of the
executor, which currently is achieved by assuming the kind of thread pool we
would use in case numThreads is not set.

Change the EvaluationContext to always require and store the parallelism (even
if the ExecutorService supplier is specified). Rename the getter to
'getParallelism' to avoid confusion with the old concept.

PiperOrigin-RevId: 239826228