Turn --experimental_use_sandboxfs into a tristate (default still "no").
This is to allow automatically using sandboxfs if it is installed in the
system (when the flag is set to "auto").

The rationale for this is to eventually allow "auto" to become the default
value for this flag so that users can trivially gain a speed-up by
installing the necessary dependency on their system.

RELNOTES: None.
PiperOrigin-RevId: 240172952