Add Asylo as an os constraint value
The Asylo project (https://asylo.dev) is using Bazel to support its toolchain.
We are providing a POSIX interface to enclave applications and operate slightly differently
from all the other POSIX operating systems. This appears to be the source of truth for what
os values there can be, so it'd be great to allow "asylo" to live among these so it's easier to
distinguish when a program is building for the Asylo platform. Ideally we wouldn't have to
state that "being on Asylo" is the same as "using the Asylo CROSSTOOL" since there is just
an ABI expectation.