Allow for empty value in --remote_http_cache
In the current bazel version attempt to disable a remote HTTP caching
set in environment is not possible. Attempt to use --remote_http_cache=
result in a NullPointerException.
This commit handles option value empty string in the same way as if
not set at all.
Also added a simple test.

This will fix the problem reported in
https://github.com/bazelbuild/bazel/issues/6719
and my request in
https://groups.google.com/forum/#!topic/bazel-dev/xDNvAbYtkkI

Closes #7650.

PiperOrigin-RevId: 240330448