Support documentation for rule families
It is often desirable to give a summary to whole family of rules in
the Build Encyclopedia, e.g., to explain who rules in this family
are to be used, or to reference other relevant documentation.

As the build-encyclopedia is generated from the doc-strings in
the individual rule classes and families are only discovered by
grouping, we have to the same route and take the information from
any rule that is willing to provide it; it is recommended to use
the "main" rule of a family (often the binary rule) for this.

Add the corresponding information to the workspace rules. While
there, also add additional links pointing to the embedded Starlark
rules.

Fixes #7798.

Change-Id: I4dbe04dfea48a060ae0d068077bc3ffc5875f4b9
PiperOrigin-RevId: 240359833