Pass FDO options to the host configuration, but only enable FDO for target configuration
This is there to allow cc_toolchain to detect that FDO is active, and to
preprocess profiles, while being analyzed in the host configuration. FDO related
options are passed to the host configuration, they are preprocessed in the host
configuration, but then used in the target configuration. Extra care is made to
not enable FDO in the host configuration (as this is the existing behavior).

https://github.com/bazelbuild/bazel/issues/6516

RELNOTES: None
PiperOrigin-RevId: 240721577