Report workspace cycles via SkylarkImportLookupKeys
A cycle in the workspace file usually means that the definition of a
repository was forgotten or comes to late in the workspace file. In
this case, we try to also report the file that was attempted to be
loaded, to help the user identify where the problem occured. So far,
we did this looking for a skykey of type AST_FILE_LOOKUP, as a Skylark
import would look for the AST. This assumption, however is invalidated
by the --incompatible_disallow_load_labels_to_cross_package_boundaries
option that first looks for the containing package of that file
(and hence produces a dependency cycle without an AST_FILE_LOOKUP).
Restore the old behaviour by obtaining the label directly from the
SKYLARK_IMPORTS_LOOKUP key in the cycle.

Closes #7826

Change-Id: Idd747caa5d8ae3bed4c3c44762c1d04717161b9b
PiperOrigin-RevId: 240771237