Add override point for cc_autoconf
This adds `overriden_labels` to the `cc_autoconf` rule. This allows to
override specific tools that bazel uses when setting up
`local_config_cc`. The benefit there is that if you want to replace
pieces of the cc toolchain, without having to create your own crosstool
entirely, this allows you to do that.

Specifically for the osx toolchain sometimes overriding flags in
response to Xcode updates, without having to wait for a bazel release is
useful.