Improve reporting about missing packages
When we report to the user that a label could not be accessed
because of of missing package, a detailed message, including the
directories searched for, would be helpful. We cannot, however,
put this information into the PackageLookupValue, as this would
increase the memory usage as individual objects then would replace
the singleton NO_BUILD_FILE_VALUE. So, in the cases the error
actually is reported to the user, query after the fact for more
information about where the package was searched for.

Fixes #7457.
Fixes #7870.

Change-Id: I652effa0bd91cc5534662baaf4c826ef79f2fd02
PiperOrigin-RevId: 240986736