remote: introduce downloadMinimal
Adds downloadMinimal() which will only download action outputs that
are strictly required and inject the remaining action metadata instead.

This change also extracts the logic of retrieving complete action result
metadata into its own method so that it can be used by both
downloadMinimal() and download().

downloadMinimal is only introduced in this CL and will be enabled in
a follow up CL.

Progress towards #6862

Closes #7867.

PiperOrigin-RevId: 240987655