Remove explicitly set target platform value and replace it by fallback.
The original value for the target platform was set at the dawn of time when we didn't have auto-calculation for it yet. Now that we do the explicit setting interferes with platform mapping which assumes that the target platform value is empty unless the user explicitly requests otherwise.

This change introduces a new flag, --target_platform_fallsback which can be used to set the fallback in case there is no matching mapping. Because it uses the old defaults no behavioral change should be observed as long as no mappings exist.

Step 5/N towards the platforms mapping functionality for https://github.com/bazelbuild/bazel/issues/6426

RELNOTES: None.
PiperOrigin-RevId: 241031610