Support optional information on the definition of a rule
External repositories are defined through the WORKSPACE file.
However, not all definitions are located directly in that file;
often repositories are definied indirectly by calling a function
imported from some bzl-file (maybe even from a repository imported
first). Therefore, record the information where such a repository
was defined, and display it in case of an error associated with
such a repository (e.g., during its fetch).

Improves on #7871.
Improves on #3892.

Change-Id: I08f72672faf672657efb082aa3a311eb428ec9a4
PiperOrigin-RevId: 241271392