C++: Bazel link/compile API
This CL partially reverts the change in https://github.com/bazelbuild/bazel/commit/769f72eb475cea0f4113ac49dd66dc3fda0608de which added link and
compile to Bazel. It replaces those methods with the new signature as discussed
in:
https://docs.google.com/document/d/1cRRdHOPTTUXBbq9Cj9hk_WLnPqsGtAoQynYd7TKBQI8/edit?usp=sharing
and adds the new method create_linking_context_from_compilation_outputs.

To keep the diff smaller, this CL doesn't do any of the implementation for
these methods yet. That will be a in a later CL.

The only new thing here are the signature changes in BazelModuleApi.java. All the code that is not in third_party is not new and has just been moved, including the tests.

RELNOTES:none
PiperOrigin-RevId: 241318078