Make testing_server.py choose the port itself
When tests need a server, our current strategy is to pick a random
unused port and tell the server to bind to that port---hoping the
port is still unused by the time the server starts up. With our range
of ports, we have a 1 in 27 000 chance that two racing processes
choose the same port to compete for; but we're trying hard enough
to actually win that race. The number of racing pairs, of course,
goes quadratic with the number of parallel test processes; the
latter has increased significantly with our switch to larger CI
machines. Also, the number of test cases has increased, thus trying
more often for a single presubmit.

Fix this flakyness by letting the server try to find an unused port
and hold to it, reporting the port found once successfully bound.

Fixes #7893.

Change-Id: I450f6e5ba70352a2d0988723911ed71a9ad0d8db
PiperOrigin-RevId: 241543075