Unbundle Android tools from Bazel's binary
Create a new package containing an external repository that the Android rules depend on at runtime.

This is an interim solution to extract dependencies out from the Bazel binary to minimize its size. The final goal is to extract all Android rules and tools out of Bazel and into rules_android and tools_android.

This change contains a new script, upload_android_tools.sh, for creating a tarball containing two runtime dependencies of the Android rules: all_android_tools_deploy.sh and ImportDepsChecker_deploy.jar. The script then versions the tarball and uploads it to GCS.

This also adds the http_archive call to the Android tools' implicit WORKSPACE to fetch the external repository. Due to this new external repository dependency, we have to add it to the distribution archive for testing as well, and renamed it from jdk_WORKSPACE to test_WORKSPACE to more accurately capture its purpose.

This is not an incompatible change and will be transparent to end users.

New Bazel size after this change, compiled with -c opt:

$ ls -lah bazel-bin/src/bazel
-r-xr-xr-x 1 jingwen primarygroup 74M Apr  1 17:44 bazel-bin/src/bazel

Fixes #1055

RELNOTES: Bazel is now ~20MiB smaller, from unbundling the Android rules' runtime dependencies.

Closes #7900.

Change-Id: Id0e115541e3e374c58fe1b1e383ad37f85690628
PiperOrigin-RevId: 241557097