Report loading cycles, even if initiated from the WORKSPACE
In the SkylarkModuleCycleReporter, also report a load cycle, if it
was initiated by a load in the WORKSPACE file (and not only if it
was a BUILD file or another .bzl file). In this way, the reporting
is always done in the same reporter, regardless of the length of
the path that lead from the WORKSPACE file to the cycle. In this
way, we also have a reporter for a cycle directly initiated by the
WORKSPACE file, avoiding an internal error.

While there, also report the path that lead to the cycle starting
at the WORKSPACE or BUILD file, not just the last entry before that.

Improves #7871.

Change-Id: Ib76932c28387f5697034cc147011a26af3bfc26c
PiperOrigin-RevId: 241727992