BES thread-safety: Reduce lock coarseness for build events while fixing some unlocked access to the pendingEvents field.
Interestingly, the crash occurs when we hit the final build event, which involves no parallelism. It is the race conditions operating on the ArrayMultimap earlier in the build which corrupts this data structure, leading to an inconsistency (an empty data structure that thinks it is non-empty).

This change yields ~5% wall-time improvement on the included benchmark.

PiperOrigin-RevId: 241790445