Properly report circular definitions of repositories
It is syntactically possible to define repositories with circular
definitions, e.g., @a might depend on a file of @b and vice versa.
These kind of circularities are discovered once the first of those
repositories is accessed (e.g., by loading a file from them). In
this case, ensure we report the cycle properly, instead of hitting
an internal error.

Improves #7871.

Change-Id: Id4227162d8c26f33e17b42242cdccafe8d02f941
PiperOrigin-RevId: 241916204