Remove third party license checking from Bazel.
This doesn't remove --incompatible_disable_third_party_license_checking
but makes it a no-op. This is so the Google version of Bazel can migrate
on its own timeframe.

Because this logic was created before Bazel existed, removing it
from Google is going to take more time. We don't want that to slow
Bazel development.

See https://github.com/bazelbuild/bazel/issues/7444.

PiperOrigin-RevId: 241946385