Add support for a new dependency in generated protos
The v1 proto package is a hodgepodge of stuff.

Of that set of stuff, generated messages only depend on a small sub-set,
which we are making accessible at //net/proto2/go:proto_gendeps.
Making generated messages depend only on what they need to allows
the proto package itself to now depend on things that generated
messages are not allowed to depend on.

RELNOTES: Generated Go protobufs now depend on //net/proto2/go:proto_gendeps instead of //net/proto2/go:proto
PiperOrigin-RevId: 242006724