Support non-fatal failures for downloads
Allow that downloads (as part of repository rules) may fail in a
way that may be handled by the rule. This is in line with the way
ctx.execute behaves. Moreover, it can be used to check if a file
is already in cache, e.g., if a download would require a credential
the user has to be asked for. To simplify pure cache probes, drop
the requirement of at least one URL being specified.

Another use case of trying downloads, while allowing failures is to
search for new versions in a free-floating definition of an archive
(to be then fixed in a resolved file).

Fixes #7635.

Change-Id: I2b3fc66b36c856077e5bd66dd5fee2af438999c9
PiperOrigin-RevId: 242089235