Extend resolved files to contain the definition location
Extend the resolved file by a an optional entry "definition_information"
for each entry providing a human-readable description of where
the repository was defined. This is useful, as resolved files are
often used to debug the list of actual repositories declared in
a workspace file, which is not always clear due to the imperative
nature of workspace files.

Change-Id: Ic31b203981b27a10cc87efeefeb8110e13658642
PiperOrigin-RevId: 242478431