Add a module reporting back where an external repository is defined
A common complaint about WORKSPACES files is that, due to their imperative
semantics, it is hard to find out where a repository is actually defined.
While we do not want to spam the overall output with all the definitions of
external repositories, report those that are involved in errors. As errors
include circular definitions, at the location of the error, the actual repository
rule is not always available, so we have to pass the information out-of-band via
the event bus.

Improves on #7871.

Change-Id: Ib6dbc22c850ab33f6149114dc5b682bc9ddbe1b6
PiperOrigin-RevId: 242634739