Fix action tracking with dynamic scheduling.
With dynamic scheduling, the same action may be in scheduling and running
states for different strategies at the same time.  We must account for
this when tracking actions and reporting on them.

To fix this, improve the ActionState structure to be more explicit about
the states it tracks and record the multiple strategies involved at each
stage.

Fixes https://github.com/bazelbuild/bazel/issues/7345.

RELNOTES: None.
PiperOrigin-RevId: 242665140