remote: implement --experimental_remote_download_outputs=toplevel
This implements the second milestone of "Remote builds without the
Bytes" and introduces the "toplevel" option for the
--experimental_remote_download_outputs flag. This mode will only
download outputs of top level targets but not outputs of intermediate
actions.

On a build of Bazel against RBE I am still seeing a 50% speed up
compared to the "all" mode.

Progress towards #6862.

Closes #7984.

PiperOrigin-RevId: 243029119