Add option to add metadata to the JSON profile.
This will currently add the build ID, the output base and a date. This data is
accessible by clicking on the button labeled with 'M' in the Chrome Trace
Viewer.

The side effect is that the format of the profile changes from JSON array to
JSON object. See
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview#

RELNOTES: None
PiperOrigin-RevId: 243071398